+++ src/context/profileContext.tsx
@@ -7,13 +7,13 @@ import {Profile} from "@/models/Profile";
 
 type ProfileContextType = {
     activeProfileInStorage: Profile | null;
-    setActiveProfileInStorage: (profileData: Partial<Profile>) => void;
+    setActiveProfileInStorage: (profileData: Partial<Profile>, triggerLoading?: boolean) => void;
     isLoading: boolean;
 };
 
 const ProfileContext = createContext<ProfileContextType | undefined>(undefined);
 
-export const ProfileProvider = ({ children }: { children: React.ReactNode }) => {
+export const ProfileProvider = ({children}: { children: React.ReactNode }) => {
     const [activeProfileInStorage, setActiveProfileInStorageState] = useState<Profile | null>(() => {
         if (typeof window !== "undefined") {
             const storedProfile = localStorage.getItem("activeProfile");
@@ -34,35 +34,40 @@ export const ProfileProvider = ({ children }: { children: React.ReactNode }) => 
         setInitialLoading(false);
     }, []);
 
-    const setActiveProfileInStorage = (profileData: Partial<Profile>) => {
-        setIsLoading(true);
+    const setActiveProfileInStorage = (profileData: Partial<Profile>, triggerLoading: boolean = true) => {
+        if (triggerLoading) {
+            setIsLoading(true);
+        }
 
         const completeProfile: Profile = {
             id: profileData.id as string,
             firstName: profileData.firstName || "",
             lastName: profileData.lastName || "",
             username: profileData.username || "",
+            status: profileData.status || 'offline',
         };
 
         localStorage.setItem("activeProfile", JSON.stringify(profileData));
         setActiveProfileInStorageState(completeProfile);
-        router.refresh();
 
-        setTimeout(() => setIsLoading(false), 3000);
+        if (triggerLoading) {
+            router.refresh();
+            setTimeout(() => setIsLoading(false), 3000);
+        }
     };
 
     if (initialLoading) {
-        return <LoadingSkeleton />;
+        return <LoadingSkeleton/>;
     }
 
     return (
-        <ProfileContext.Provider value={{ activeProfileInStorage, setActiveProfileInStorage, isLoading }}>
+        <ProfileContext.Provider value={{activeProfileInStorage, setActiveProfileInStorage, isLoading}}>
             {children}
         </ProfileContext.Provider>
     );
 };
 
-export const useProfile = () => {
+export const useProfileContext = () => {
     const context = React.useContext(ProfileContext);
     if (context === undefined) {
         throw new Error("useProfile must be used within a ProfileProvider");